merge traces with randomized time periods
Each tracefile is followed by a numeric parameter that is treated
as the target average number of instructions that it will execute
before the ARQ switches to the other.  The sojourn times are
generated by an exponential distribution, so the exponential
parameter is exactly the inverse (i.e. 1/x) of the given parameter.

Tested as follows:
% ./tracer/merge_traces /dev/stdout <(xz -dc dpc3_traces/605.*) 1000000 <(xz -dc dpc3_traces/641.*) 10000 | $CHAMPSIM -stdin